Move path calculations to separate file
